refactor : Components and hooks seperated
Created several components and hooks due to DRY - SOLID principles
